rename sections to atomic-layout-section-utils
https://coveord.atlassian.net/browse/KIT-5096
