+++ packages/atomic/src/components/common/atomic-layout-section/atomic-layout-section-utils.spec.ts
@@ -1,8 +1,8 @@
 import {describe, expect, it} from 'vitest';
-import type {Section} from './sections';
-import {findSection, sectionSelector} from './sections';
+import type {Section} from './atomic-layout-section-utils';
+import {findSection, sectionSelector} from './atomic-layout-section-utils';
 
-describe('sections', () => {
+describe('atomic-layout-section-utils', () => {
   describe('#sectionSelector', () => {
     it('should return correct selector for each section', () => {
       const sections: Section[] = [
