+++ packages/atomic/src/components/search/atomic-refine-modal/atomic-refine-modal.tsx
@@ -26,7 +26,7 @@ import {
 } from '../../../utils/initialization-utils';
 import {shouldDisplayOnCurrentTab} from '../../../utils/tab-utils';
 import {sortByDocumentPosition} from '../../../utils/utils';
-import {findSection} from '../../common/atomic-layout-section/sections';
+import {findSection} from '../../common/atomic-layout-section/atomic-layout-section-utils';
 import {popoverClass} from '../../common/facets/popover/popover-type';
 import {
   BaseFacetElement,
